java_jni/refactoring: move class self type handling to separate file
